refactor: separate BSE module and input items
