+++ source/source_esolver/esolver_lr_lcao_bse.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "esolver_lr_lcao_tddft.h"
-#include "source_lcao/module_lr/bse/molecular_lri.h"
+#include "source_lcao/module_bse/molecular_lri.h"
 
 namespace ModuleESolver
 {
