+++ source/source_io/module_parameter/read_input.cpp
@@ -173,6 +173,7 @@ ReadInput::ReadInput(const int& rank)
     this->item_rt_tddft();
     this->item_tdofdft();
     this->item_lr_tddft();
+    this->item_bse();
     this->item_output();
     this->item_postprocess();
     this->item_model();
