+++ source/source_lcao/module_bse/bse_util.cpp
