+++ source/source_lcao/module_bse/molecular_lri.h
@@ -12,7 +12,7 @@
 #include "source_cell/klist.h"
 #include "source_hamilt/module_xc/exx_info_ri.h"
 #include "source_lcao/module_ri/lri_cv_tools.h"
-#include "source_lcao/module_lr/bse/bse_util.h"
+#include "source_lcao/module_bse/bse_util.h"
 #include "source_lcao/module_lr/utils/lr_util.h"
 #include "source_lcao/module_lr/utils/lr_util_print.h"
 #include "source_lcao/module_lr/ao_to_mo_transformer/ao_to_mo.h"
