+++ source/source_lcao/module_bse/test/CMakeLists.txt
@@ -5,7 +5,7 @@ if(ENABLE_ELPA)
 endif()
 
 AddTest(
-  TARGET MODULE_LR_bse_solver_test
+  TARGET MODULE_BSE_solver_test
   LIBS ${bse_solver_test_libs}
   SOURCES hamilt_bse_solver_test.cpp ../hamilt_bse_solver.cpp
 )
